+++ src/objects/game_object.h
@@ -40,7 +40,7 @@ namespace sun_magic {
 			z_ = z;
 		}
 
-		virtual GameObject* UpdateFocus(float mouse_x, float mouse_y, sf::Vector2i& abs_pos) {
+		virtual GameObject* UpdateFocus(int mouse_x, int mouse_y, sf::Vector2f& abs_pos) {
 			abs_pos.x += rect_.left;
 			abs_pos.y += rect_.top;
 			return this;
